General cleanup of the rxp reader
This patch removes the inclination corrector entirely, since it was
highly experimental and shouldn't really be used by the general public.
It also splits the reader and the pointcloud into two files for better
compartmentalization, and makes some basic type fixes (unique_ptr
instead of shared_ptr, point_count_t instead of something else, etc).
